+++ Samples~/Scripts/Dummy.cs
@@ -4,6 +4,6 @@ namespace SaintsField.Samples.Scripts
 {
     public class Dummy: MonoBehaviour
     {
-
+        public int number;
     }
 }